# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API GPE",
    "version": "1.0.0",
    "description": "Documentação da API do sistema GPE desenvolvida em Node.js com Express e MySQL."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/turmas": {
      "get": {
        "tags": ["Gestores"],
        "summary": "Lista todas as turmas",
        "responses": {
          "200": {
            "description": "Lista de turmas retornada com sucesso"
          },
          "500": {
            "description": "Erro ao listar turmas"
          }
        }
      },
      "post": {
        "tags": ["Gestores"],
        "summary": "Cria uma nova turma",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "1º Ano A"
                  },
                  "ano": {
                    "type": "integer",
                    "example": 2026
                  },
                  "turno": {
                    "type": "string",
                    "example": "matutino"
                  }
                },
                "required": ["nome", "ano"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Turma criada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro ao criar turma"
          }
        }
      }
    },
    "/api/turmas/{id}": {
      "get": {
        "tags": ["Gestores"],
        "summary": "Busca uma turma por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turma encontrada"
          },
          "404": {
            "description": "Turma não encontrada"
          },
          "500": {
            "description": "Erro ao buscar turma"
          }
        }
      },
      "put": {
        "tags": ["Gestores"],
        "summary": "Atualiza uma turma",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "2º Ano B"
                  },
                  "ano": {
                    "type": "integer",
                    "example": 2026
                  },
                  "turno": {
                    "type": "string",
                    "example": "vespertino"
                  },
                  "ativa": {
                    "type": "integer",
                    "example": 1
                  }
                },
                "required": ["nome", "ano"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Turma atualizada com sucesso"
          },
          "404": {
            "description": "Turma não encontrada"
          },
          "500": {
            "description": "Erro ao atualizar turma"
          }
        }
      },
      "delete": {
        "tags": ["Gestores"],
        "summary": "Exclui uma turma",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turma excluída com sucesso"
          },
          "404": {
            "description": "Turma não encontrada"
          },
          "500": {
            "description": "Erro ao excluir turma"
          }
        }
      }
    },
    "/api/turmas/{id}/alunos": {
      "get": {
        "tags": ["Gestores"],
        "summary": "Lista os alunos de uma turma",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alunos da turma retornada com sucesso"
          },
          "500": {
            "description": "Erro ao buscar alunos da turma"
          }
        }
      }
    },
    "/api/turmas/{id}/professores": {
      "post": {
        "tags": ["Gestores"],
        "summary": "Vincula um professor a uma turma",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "professor_id": {
                    "type": "integer",
                    "example": 3
                  }
                },
                "required": ["professor_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor vinculado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro ao vincular professor"
          }
        }
      }
    },
    "/api/turmas/{id}/professores/{professor_id}": {
      "delete": {
        "tags": ["Gestores"],
        "summary": "Remove o vínculo de um professor com uma turma",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "professor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor removido da turma com sucesso"
          },
          "404": {
            "description": "Vínculo não encontrado"
          },
          "500": {
            "description": "Erro ao remover professor da turma"
          }
        }
      }
    },
...
    "/api/alunos/{id}/turma": {
      "patch": {
        "tags": ["Alunos"],
        "summary": "Atualiza a turma de um aluno",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "turma_id": {
                    "type": "integer",
                    "example": 2
                  }
                },
                "required": ["turma_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Turma do aluno atualizada com sucesso"
          },
          "404": {
            "description": "Aluno não encontrado"
          },
          "500": {
            "description": "Erro ao atualizar turma do aluno"
          }
        }
      }
    },
...
  }
}